# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for dev.",
        "title": "dtle API Docs",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v2/mysql/columns": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "list columns of mysql source instance.",
                "tags": [
                    "mysql"
                ],
                "operationId": "ListMysqlColumnsV2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mysql host",
                        "name": "mysql_host",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql port",
                        "name": "mysql_port",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql user",
                        "name": "mysql_user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql password",
                        "name": "mysql_password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql schema",
                        "name": "mysql_schema",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql table",
                        "name": "mysql_table",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql character set",
                        "name": "mysql_character_set",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "indecate that mysql password is encrypted or not",
                        "name": "is_mysql_password_encrypted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListColumnsRespV2"
                        }
                    }
                }
            }
        },
        "/v2/mysql/instance_connection": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "connect to  mysql instance.",
                "tags": [
                    "mysql"
                ],
                "operationId": "MySQLConnectionV2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mysql host",
                        "name": "mysql_host",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql port",
                        "name": "mysql_port",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql user",
                        "name": "mysql_user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql password",
                        "name": "mysql_password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "indecate that mysql password is encrypted or not",
                        "name": "is_mysql_password_encrypted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ConnectionRespV2"
                        }
                    }
                }
            }
        },
        "/v2/mysql/schemas": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "list schemas of mysql source instance.",
                "tags": [
                    "mysql"
                ],
                "operationId": "ListMysqlSchemasV2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "mysql host",
                        "name": "mysql_host",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql port",
                        "name": "mysql_port",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql user",
                        "name": "mysql_user",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql password",
                        "name": "mysql_password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mysql character set",
                        "name": "mysql_character_set",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "indecate that mysql password is encrypted or not",
                        "name": "is_mysql_password_encrypted",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ListSchemasRespV2"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ConnectionRespV2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ListColumnsRespV2": {
            "type": "object",
            "properties": {
                "columns": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ListSchemasRespV2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "schemas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SchemaItem"
                    }
                }
            }
        },
...
        "models.SchemaItem": {
            "type": "object",
            "properties": {
                "schema_name": {
                    "type": "string"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TableItem"
                    }
                }
            }
        },
...
        "models.TableItem": {
            "type": "object",
            "properties": {
                "table_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}